+++ cogs/modmail.py
@@ -367,7 +367,6 @@ class Modmail:
         await session.run()
         
     @commands.command()
-    @trigger_typing
     async def reply(self, ctx, *, msg=''):
         """Reply to users using this command.
 
@@ -376,15 +375,25 @@ class Modmail:
         ctx.message.content = msg
         thread = await self.bot.threads.find(channel=ctx.channel)
         if thread:
+            await ctx.trigger_typing()
             await thread.reply(ctx.message)
     
     @commands.command()
-    @trigger_typing
+    async def anonreply(self, ctx, *, msg=''):
+        ctx.message.content = msg
+        thread = await self.bot.threads.find(channel=ctx.channel)
+        if thread:
+            await ctx.trigger_typing()
+            await thread.reply(ctx.message, anonymous=True)
+    
+    @commands.command()
     async def note(self, ctx, *, msg=''):
         """Take a note about the current thread, useful for noting context."""
         ctx.message.content = msg 
         thread = await self.bot.threads.find(channel=ctx.channel)
+
         if thread:
+            await ctx.trigger_typing()
             await thread.note(ctx.message)
 
     @commands.command()